ops functions to separate file
